_archive/ThreadToHitRemix/requirements.txt
anthropic>=0.39.0
accelerate>=0.32.1
audiocraft>=1.2.0
librosa>=0.10.2
soundfile>=0.12.1
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
requests>=2.32.0
scikit-learn>=1.5.0
scipy>=1.11.0
pyloudnorm>=0.1.1
